End of WIP in tests, all green
