+++ tests/Serializer/RepresentationSerializerTest.php
@@ -25,6 +25,21 @@ class RepresentationSerializerTest extends AbstractSerializerTest
     }
 
     /**
+     * @param bool $serializeAllObjects
+     * @dataProvider serializeAllObjectsProvider
+     */
+    public function testArraysAreArrays($serializeAllObjects)
+    {
+        $serializer = $this->getSerializerUnderTest();
+        if ($serializeAllObjects) {
+            $serializer->setAllObjectSerialize(true);
+        }
+        $input = [1, 2, 3];
+        $result = $this->invokeSerialization($serializer, $input);
+        $this->assertSame(['1', '2', '3'], $result);
+    }
+
+    /**
      * @param bool $serialize_all_objects
      * @dataProvider serializeAllObjectsProvider
      */
